Modified layer wise prune enable bool to layer specific scaling

+++ include/darknet.h
@@ -153,7 +153,7 @@ struct layer{
     int steps;
     int hidden;
     int truth;
-    int prune;
+    float prune;
     float smooth;
     float dot;
     float angle;
